Almost finished working on new expression codgen. almost works except arrays

+++ test_files/ref_deref_tests/simple_ref.c
@@ -1,5 +1,5 @@
-int main() {
-    int x = 69;
-    int* y = &x;
+long main() {
+    int  x = 69;
+    int * y = &x;
     return *y;
 }
